refactor: extract more stages from inbound-message-handler to meet <600 line target
